Accept workspace relative path for erlangPath
Erlang/OTP may not be installed on the host but prepared in a project
specific directory. This directory could depend on where the workspace
(e.g. the git repository) is.
Accept workspace relative path for erlang.erlangPath and resolve it
based on the workspace root directory.

Check if erlang.erlangPath is a Windows absolute path and discard it on
non-Windows systems. For example in remote development when user
settings has a Windows path but the remote host is Linux.