Follow up to #183 to fix #182 for Windows users. (#184)
* Ignore rebar.lock and *.vsix files.

Signed-off-by: Hemanth V. Alluri <hdrive1999@gmail.com>

* Fix pgourlain#182 in a more cross-platform way.

The changes in pgourlain#183 have only partially fixed the underlying
issue. Now Unix users won't face any problems with spaces in the path
to Rebar3, but Windows users still will. The reason for this is that
in GenericShell.ts, in the LaunchProcess method, we use:

this.childProcess = spawn(processName, args,
    { cwd: startDir, shell: true, stdio: 'pipe', env : childEnv });

On Windows, it turns out that the shell does not treat ' and " as the
same. ' is treated as a normal character but " is treated as an actual
quotation character (which is what we want).

Signed-off-by: Hemanth V. Alluri <hdrive1999@gmail.com>