Fix parsing record match spec and go to record
Ignore non-atom record fields during finding record name or record field
under the cursor. If record is used in a match specification then
underscore (_) is legal to use instead of a real field name.
For example: #rec{a = 1, _ = '_'}.

Accept file path (URI) of the record definition in both string and
binary format when go to record definition. I am not sure exactly when
it changed and why but it seems 'epp' returns file names in record
attributes as binaries instead of strings for a while.