Add new settings to support development
Add setting `erlang.erlangArgs` for extra command line arguments passed
to the Erlang counterpart. For example `-S 4:4` to limit the CPU usage
on large systems. It is an experimental feature and not used by default.

Add setting `erlang.erlangDistributedNode` to start the Erlang
counterpart in a distributed Erlang node that makes possible to
connect to it and investigate various issues. (E.g. huge memory usage
addressed in the next commit.)