Navigation updates
Go to definition of local functions is improved. Always jump to local
function definition regardless of there are multiple source modules with
the same name (see commit 8f4eac97).

Go to definition of macro `?MODULE` is shortcut to the current source
module file, as `MODULE` is not explicitly defined in source code.